fix: use parseFloat instead of parse int

+++ src/controllers/finance/takeLoan.ts
@@ -18,7 +18,7 @@ const takeLoan: RequestHandler = async (req, res, next) => {
   if (!amount) {
     return next(new AppError('amount is required', 400));
   }
-  const _amount = parseInt(amount);
+  const _amount = parseFloat(amount);
 
   if (!_amount || isNaN(_amount)) {
     return next(new AppError('invalid value for amount', 400));
